Updated .cproj configuration
The fucking shit in the folders and resources mixed in with the code is maddening.
I found a solution though, and put a specific version for CodeAnalysis so that it wouldn't fill the whole folder with its language packs.
I also rewrote the command to copy the resources folder to the build.